선경님_0603PR입니다!

+++ src/main/java/com/team/berp/domain/OrderLineItem.java
@@ -32,7 +32,7 @@ public class OrderLineItem {
     private Warehouse warehouse;
 
     @Column
-    private Long unitQty;
+    private Integer unitQty;
 
     @Column
     private Long unitPrice;
